Move remaining tendermint-types into one file
