--- packages/iov-tendermint-types/src/network.ts
@@ -1,5 +1,0 @@
-import { As } from "type-tagger";
-
-// ChainId is used to differentiate a blockchain
-// should be alphanumeric or -_/ and unique
-export type ChainId = string & As<"chain-id">;
