--- packages/iov-tendermint-types/types/network.d.ts
@@ -1,2 +1,0 @@
-import { As } from "type-tagger";
-export declare type ChainId = string & As<"chain-id">;
