--- packages/iov-tendermint-types/types/transactions.d.ts
@@ -1,4 +1,0 @@
-import { As } from "type-tagger";
-export declare type SignatureBytes = Uint8Array & As<"signature">;
-export declare type PostableBytes = Uint8Array & As<"postable">;
-export declare type TxId = Uint8Array & As<"txid">;
